Reorganise visitor classes and files
